src/requirements.txt
fastapi[all]
aiosqlite
tortoise-orm[accel]